+++ internal_filesystem/lib/mpos/board/fri3d_2026.py
@@ -23,7 +23,7 @@ import drivers.display.st7789 as st7789
 
 import mpos.ui
 import mpos.ui.focus_direction
-from mpos import InputManager
+from mpos import InputManager, IRManager
 
 spi_bus = SPI.Bus(
     host=2,
@@ -195,6 +195,9 @@ InputManager.register_indev(indev)
 
 import mpos.sdcard
 mpos.sdcard.init(spi_bus=spi_bus, cs_pin=14)
+
+IRManager.txPin = Pin(10, Pin.OUT)
+IRManager.rxPin = Pin(11, Pin.IN)
 
 # === AUDIO HARDWARE ===
 from mpos import AudioManager