+++ CHANGELOG.md
@@ -6,9 +6,13 @@ Add changes that have been made to the code but haven't made it into a release h
 Frameworks:
 - AppManager: support .mpk/.zip files with compression and an unnecessary top-level directory
 - Add new LoRaManager framework
+- Add new IRManager framework
 
 OS:
-- Add Infra Red (IR) RMT drivers for ESP32
+- LilyGo T-Watch S3 Plus: add support for IR Remote app
+- Fri3d 2024: add support for IR remote app
+- Fri3d 2026: add support for IR remote app
+
 
 0.9.2
 =====